Similar to find_missing_runs, but excised from prep_run_save_ewrs to take the same arguments

Similar to find_missing_runs, but excised from prep_run_save_ewrs to take the same arguments

Usage

check_missing_runs(
  hydro_dir,
  output_parent_dir,
  output_subdir = "",
  scenarios = NULL,
  file_search = NULL,
  model_format = "Standard time-series",
  outputType = "none",
  scenarios_from = "directory"
)

Arguments

hydro_dir

Directory containing hydrographs. Can be an outer directory, e.g. hydrographs that splits into scenario subdirs, or can be a single scenario subdir.

output_parent_dir

parent directory for the outputs. Can be anything, but two typical cases:

The directory containing hydro_dir, which puts the module_outputs at the same level as the hydrographs

If running in batches for single scenarios, may be hydro_dir, which just puts the module_outputs in hydro_dir

output_subdir

a sub-directory for the outputs, if for example we want module_output/EWR/V1 and module_output/EWR/V2

scenarios

NULL (default) or named list.

NULL- finds scenario names by parsing directory names in hydro_dir. If no internal directories, just stays in hydro_dir. This captures the two typical situations discussed for output_parent_dir. If there are other directories in hydro_dir that do not contain hydrological scenarios, should use a character vector.

named list of paths to files. names become scenario names, paths should be relative to hydro_dir. This allows unusual directory structures.

file_search

character, regex for additional limitations on filenames. Useful to run a subset of scenarios or if several files have the extension defined by model_format, but only some are hydrographs.

model_format

see EWR tool. One of:

'Standard time-series': (default, among other things accepts a csv with a Date column followed by gauge columns, with _flow or _level appended to the gauge number)

'IQQM - netcdf': in development, finds all netcdf files in hydro_dir. Should also work when hydro_dir is a .zip with netcdfs inside

'ten thousand year': old default (IQQM - NSW 10,000 years), works nearly the same as standard time-series

'All Bigmod': previously 'Bigmod - MDBA'

'Source - NSW (res.csv)'

outputType

list of strings or character vector defining what to save to disk. One or more of:

'none' (default), do not save outputs- ignored if in a list with others

'summary',

'yearly',

'all_events',

'all_successful_events',

'all_interEvents', # Does not work with EWR 2.0 tool

'all_successful_interEvents'

scenarios_from

character, default 'directory' gets scenario names from directory names. If anything else, gets them from filenames (safest). Expect additional options in future, e.g from metadata.

Value

character vector of missing filepaths