Fixes potential issue with non-unique scenario names, which may involve file renaming

'scenario' here gets the csv name. That happens inside the EWR tool- the _get_file_names function does a .split('/') on the filepath to name the scenario. This causes issues when we have csvs per gauge, each with just the gaugename. In that case the EWR tool assigns them the gaugename as the scenario. So we need to ensure that doesn't happen. I think probably the best way to do this is to encourage the use of single files per scenario, and if not, filenames that have the scenario name in them with the gauge name. This function tests if there's an issue, tries to fix it and informs or warns the user to name files better

Usage

fix_file_scenarios(hydro_paths, scenarios)

Arguments

hydro_paths

paths to hydrographs, typically from find_scenario_paths()

scenarios

names of the scenarios, useful for cross-checking

Value

paths to hydrographs, possibly fixed