Pairs gauge numbers with geographic locations to make an sf

Typically used for EWR outputs, but anything with a gauge column with gauge numbers as characters will work.

Usage

join_to_geo(data_df, spatial_locs, whichcrs = 4283)

Arguments

data_df

A tibble or dataframe with a gauge column with gauge numbers as characters

spatial_locs

sf or path to the csv with gauge numbers paired with lat-long.

whichcrs

desired crs of the output. input crs assumed to be WGS84, EPSG 4326

Value

an sf dataframe with gauge locations as points in the geometry