Set up the output directory structure based on parent directory and module at parent_dir/module_output/MODULE_NAME/subdir

Set up the output directory structure based on parent directory and module at parent_dir/module_output/MODULE_NAME/subdir

Usage

make_output_dir(
  parent_dir,
  scenarios,
  module_name = "EWR",
  subdir = "",
  ewr_outtypes = c("summary", "yearly")
)

Arguments

parent_dir

parent directory- typically for a project, but could be a single scenario

scenarios

names of the scenarios (character vector)

module_name

default 'EWR', sets up different directories for different modules

subdir

character, default "" for none, but if included, makes a sub-directory in parent_dir/module_output/MODULE_NAME

ewr_outtypes

character vector, names of EWR outputs to return. Options include

'summary': outputs summarised to the period

'yearly': outputs summarised to year

'all_events'

'all_successful_events'

'all_interEvents'

'all_successful_interEvents'

Value

path to output directory