Make a bar plot (qualitative x, usually)

main function to make a standard bar plot with ggplot2::geom_col()

Usage

plot_bar(prepped, x_col, x_lab, outcome_lab, position, transoutcome)

Arguments

prepped

prepared data from plot_data_prep() |> plot_style_prep()

x_col

character, name of column for the x-axis. Default 'scenario'

x_lab

character, default x_col, allows changing x-label

outcome_lab

character, default outcome_col, allows changing label of the outcome_col (y-axis or color/fill)

position

character or position function, position arguments from ggplot2::geom_col() and ggplot2::geom_point() (depending on plot type), to change from stacked to dodged bars or jitter points. Can be character, e.g. 'jitter' or a function, e.g. ggplot2::position_jitter(width = 0.1, height = 0)

transoutcome

transformation for outcome as in ggplot2::scale_y_continuous() or ggplot2::scale_fill_continuous(). Default transoutcome = 'identity' just uses the data. Most common change likely transoutcome = 'log10

Value

ggplot object