Make a heatmap

Make a heatmap

Usage

plot_heatmap(
  prepped,
  x_col,
  x_lab,
  y_col,
  y_lab,
  outcome_lab,
  transoutcome,
  transx,
  transy,
  contour_arglist,
  xtype,
  ytype,
  setLimits,
  base_list
)

Arguments

prepped

prepared data from plot_data_prep() |> plot_style_prep()

x_col

character, name of column for the x-axis. Default 'scenario'

x_lab

character, default x_col, allows changing x-label

y_col

character, default outcome_col. Allows separately specifying y-axis from outcome (e.g. maps, heatmaps, networks)

y_lab

character, default y_col, allows changing y-label

outcome_lab

character, default outcome_col, allows changing label of the outcome_col (y-axis or color/fill)

transoutcome

transformation for outcome as in ggplot2::scale_y_continuous() or ggplot2::scale_fill_continuous(). Default transoutcome = 'identity' just uses the data. Most common change likely transoutcome = 'log10

transx

transformation for x axis as in ggplot2::scale_x_continuous(). Default transx = 'identity' just uses the data. Most common change likely transx = 'log10.

transy

transformation of y-axis, if the outcome is not y, when transoutcome should be used. Default 'identity'. Ignored if outcome_col is on the y-axis, because transoutcome should be used.

contour_arglist

default NULL

NULL builds a heatmap with ggplot2::geom_tile(),

list(interpolate = TRUE) builds an interpolated heatmap with ggplot2::geom_raster() with interpolate = TRUE

list() builds a ggplot2::geom_contour() with all defaults

a named list with names other than 'interpolate' passes those as arguments to ggplot2::geom_contour()

xtype

character, 'qual', 'numeric' or 'date' (which covers several different date types). Determines how axes are handled

ytype

character, 'qual', 'numeric' or 'date' (which covers several different date types). Determines how axes are handled

setLimits

sets user-supplied color/fill limits or midpoints for maps, heatmaps, and networks, or y limits for other plots. Also sets underlay and overlay limits for consistency.

NULL (default) use internally-calculated limits

length-1 numeric, sets midpoint, limits stay auto-calculated. V. useful for divergent

length-2 numeric, sets lower and upper limits

length-3 numeric, sets limits and midpoint c(lower, mid, upper), with caveat that one limit may be ignored (must be symmetrical about the midpoint)

if base_list is 'difference' or 'relative', midpoints are auto-set at 0 or 1 (respectively) unless otherwise specified here.

base_list

NULL (default) or named list of arguments to baseline_compare();

base_lev

comp_fun

group_cols plot_data_prep() handles zero_adjust, and other arguments are inferred or not supported If comp_fun is 'difference' or 'relative', midpoint auto-set at 0 or 1 respectively unless otherwise specified with setLimits

Value

ggplot object