Title

Plots a standard map, with potentially underlay and overlays, specified by underlay_list and overlay_list. These are recursed through the plot_*_prep() functions so the transes and other changes are consistent across the layers. There are some inherent limitations about color/fill and points/polygons due to masking and use of multiple color/fill ramps

Usage

plot_map(
  prepped,
  underlay_list,
  overlay_list,
  map_outlinecolor = "grey35",
  outcome_lab,
  facet_wrapper,
  facet_row,
  facet_col,
  sceneorder,
  transoutcome,
  setLimits,
  base_list
)

Arguments

prepped

prepared data from plot_data_prep() |> plot_style_prep()

underlay_list

default NULL, otherwise named list (or list of named lists for multiple underlay levels) of arguments to plot a map underlying the main map data. Only required value is either 'underlay' or 'underover' (interchangeable), defining the data to be plotted. Other arguments as for the main plot. Typical examples

underlay_list = 'basin' simply plot the basin outline

underlay_list = list(underlay = 'basin', pal_list = 'azure') same, but filled

underlay_list = list(underlay = outputs_on_sdl_scale, outcome_col = ewr_achieved, pal_list = 'scico::oslo') plot the ewr outcomes in the sdls A named palette (e.g. variation in color/fill) only works if the main data is not the same type- we can't use different palettes for underlay fill and main data fill, for example, but can if the underlay is fill (polygons) and the main data is points.

overlay_list

as underlay_list, but first item might be named "overlay or "underover"

map_outlinecolor

color specification for the outline of filled areas on maps. Default 'grey35' seems to be the sf default. NA removes the outline.

outcome_lab

character, default outcome_col, allows changing label of the outcome_col (y-axis or color/fill)

facet_wrapper

NULL (default) or character for column to use for facet_wrap

facet_row

NULL (default) or character for facet row. Can be '.' to have one row and column-facets

facet_col

NULL (default) or character for facet column. Can be '.' to have one column and row-facets

sceneorder

Default NULL, otherwise, character or factor giving the order to present scenario levels

transoutcome

transformation for outcome as in ggplot2::scale_y_continuous() or ggplot2::scale_fill_continuous(). Default transoutcome = 'identity' just uses the data. Most common change likely transoutcome = 'log10

setLimits

sets user-supplied color/fill limits or midpoints for maps, heatmaps, and networks, or y limits for other plots. Also sets underlay and overlay limits for consistency.

NULL (default) use internally-calculated limits

length-1 numeric, sets midpoint, limits stay auto-calculated. V. useful for divergent

length-2 numeric, sets lower and upper limits

length-3 numeric, sets limits and midpoint c(lower, mid, upper), with caveat that one limit may be ignored (must be symmetrical about the midpoint)

if base_list is 'difference' or 'relative', midpoints are auto-set at 0 or 1 (respectively) unless otherwise specified here.

base_list

NULL (default) or named list of arguments to baseline_compare();

base_lev

comp_fun

group_cols plot_data_prep() handles zero_adjust, and other arguments are inferred or not supported If comp_fun is 'difference' or 'relative', midpoint auto-set at 0 or 1 respectively unless otherwise specified with setLimits

Value

ggplot object