Some data preparation, checking, and type-finding based on the type of data and the way we want plots to look

Some data preparation, checking, and type-finding based on the type of data and the way we want plots to look

Usage

plot_style_prep(
  prepped,
  colorset,
  colorgroups,
  pal_list,
  pal_direction,
  transoutcome,
  transx,
  point_group
)

Arguments

prepped

a prepped data list returned by plot_data_prep()

colorset

character, column name to use for color or fill

colorgroups

character, column name for grouping colorset to allow multiple palettes. Default NULL just uses colorset to define color

pal_list

list of palettes for defining colors for colorset. Should be length of colorgroups

pal_direction

vector of length pal_list, either 1 (default) or -1 (reversed) direction of the palettes

transoutcome

transformation for outcome as in ggplot2::scale_y_continuous() or ggplot2::scale_fill_continuous(). Default transoutcome = 'identity' just uses the data. Most common change likely transoutcome = 'log10

transx

transformation for x axis as in ggplot2::scale_x_continuous(). Default transx = 'identity' just uses the data. Most common change likely transx = 'log10.

point_group

character, column to add additional point groupings to, e.g. if plotting color by an environmental group, but need separate lines/points for each environmental objective.

Value

a prepped data list with more information needed to make the plots