Joins spatial from and to sf objects with needed data prep for the aggregate functions

This ensures the polygons have the same crs, are valid, optimises the sf::st_intersection, calculates an area column for future use, and drops geometry while retaining the polyID from the from_geo as a grouper for further aggregation and later re-joining to the geometry

Usage

spatial_joiner(from_geo, to_geo, whichcrs)

Arguments

from_geo

sf of the input data to be aggregated

to_geo

sf of the polygons or multipolygons to aggregate into

whichcrs

desired crs. Either the numeric EPSG or a crs definition, as from sf::st_crs()

Value

tibble (not sf) with a polyID column mapping each row to it's relevant polygon in polyID and an area column for the area of each polygon resulting from the intersection of from_geo and to_geo. Rows do not match to_geo because to_geo is intersected with from_geo, and so each row is a unique combination of the two (from_geo rows may be duplicated across several to_geo polygons). Each of these unique polygons has the area calculated and added as a column. The geometry is dropped for speed, and expected to add back in later with a dplyr::left_join to to_geo after aggregation or other processing.